PAYARA-895 some clean up and extracting ClusteredCDIEventBus into an interface.
