+++ appserver/payara-appserver-modules/payara-micro-cdi/src/main/resources/META-INF/services/javax.enterprise.inject.spi.Extension
@@ -1,1 +1,1 @@
-fish.payara.micro.cdi.PayaraMicroCDIExtension
+fish.payara.micro.cdi.extension.PayaraMicroCDIExtension
